eliminate python-ffmpeg dependency
since it isn't available in Fedora 42, anymore,
and upstream seems to be inactive since a few years.

Also, cutbynoise's use of ffmpeg is relatively tame
such that it arguably doesn't profit from advanced
abstractions (such as when constructing advanced filter pipelines),
anyways.

Last but not least, keeping the number of depencies low is nice, as well.

cf.
https://github.com/kkroening/ffmpeg-python
https://src.fedoraproject.org/rpms/python-ffmpeg-python